Retry when encountering negative execution times
Benchmarking a system with a broken clock is probably a questionable
idea to begin with ... but from my testing it seems like the stats
coming out of docker for mac are still useful-ish, so it's probably
better to report them rather than crashing hard. Users can tell
something bad is happening using the new errors counter.